+++ nImO/nImOchannelName.h
@@ -96,8 +96,8 @@ namespace nImO
             generatePath
                 (const std::string &    base,
                  const bool             forOutput,
-                 const uint16_t         numChannels,
-                 const uint16_t         channelNumber,
+                 const int64_t          numChannels,
+                 const int64_t          channelNumber,
                  std::string &          path);
 
             /*! @brief Return the name in standard form.
